Extract logic for calculating results from rest of cache logic
